% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winflashr-package.R
\docType{package}
\name{winflashr}
\alias{winflashr}
\alias{winflashr-package}
\title{Blink Rstudio window in Windows taskbar}
\description{
The package just contains one function \code{\link{winflash}} which can be
used to make Rstudio window blink (caption and taskbar) in the end of a long
running command. Windows only.
}
\examples{
# Blink window 10 times
winflash(10)

}
\author{
Yuri Pimenov <pimenoff@gmail.com>
}

